*** empty log message ***
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@19176 72102866-910b-0410-8b05-ffd578937521

+++ FSNode/FSNBrowserCell.h
@@ -50,7 +50,7 @@ typedef NSString *(*cutIMP)(id, SEL, id, float);
   NSString *selectionTitle; 
   
   NSImage *icon;
-  float icnsize;
+  int icnsize;
   
   NSBezierPath *highlightPath;
   NSRect hlightRect;
